[Clang] Don't disable C++ exceptions when using a Kitsune front end.
Disabling C++ exceptions makes it hard to use a Kitusne front end to compile large applications.  In such applications, it's relatively easy to disable exceptions when compiling individual source files that contain Kitsune constructs and no need for exceptions.  But other sources in the same application may require exceptions, and it's harder to change which compiler front end to use for different source files.